requirements.txt
pandas
requests
openpyxl
python-dotenv
fastapi
uvicorn
python-multipart
oracledb
firebase-admin
